2021/06/07-23:09

+++ server/routes/projectsRouter.js
@@ -1,23 +1,171 @@
 const express = require("express");
 const { Project, validateProject } = require("../models/project");
+const { User } = require("../models/user");
 const auth = require("../middleware/auth");
 const router = express.Router();
 const fs = require("fs");
 const path = require("path");
 const multer = require("multer");
 
+router.post("/", auth, async (req, res) => {
+  if (req.user && req.user.isAdmin) {
+    const { name } = req.body;
+    const folder = name.toLowerCase();
+
+    const makeDir = (folder, key) => {
+      const image = path.basename(req.body[key]);
+      return `/images/projects/${folder}/${image}`;
+    };
+
+    const checkName = name => name.replace(/-/g, " ");
+    const checkReq = reqItem => (reqItem ? reqItem : "לא מוגדר");
+    const checkUrl = url =>
+      url ? makeDir(folder, req.body[key]()) : "/images/logo/logo_black.png";
+    const checkPath = path => (path ? path.basename(path) : "לא מוגדר");
+
+    let project = {
+      name: checkName(req.body.name),
+      year: req.body.year,
+      size: req.body.size,
+      category: req.body.category,
+      address: {
+        country: req.body.country,
+        city: req.body.city,
+        street: req.body.street,
+        houseNumber: req.body.houseNumber,
+        zip: checkReq(req.body.zip),
+      },
+      description: checkReq(req.body.description),
+      files: {
+        contracts: [
+          {
+            name: checkPath(req.body.contract),
+            url: checkUrl(req.body.contract),
+            remarks: "כאן ניתן לכתוב הערות",
+          },
+        ],
+        licensing: [
+          {
+            name: checkPath(req.body.licensing),
+            url: checkUrl(req.body.licensing),
+            remarks: "כאן ניתן לכתוב הערות",
+          },
+        ],
+        experts: [
+          {
+            firstName: checkReq(req.body.expertFirstName),
+            lastName: checkReq(req.body.expertLastName),
+            phone: checkReq(req.body.expertPhone),
+            files: [
+              {
+                name: checkPath(req.body.expertFile),
+                url: checkUrl(req.body.expertFile),
+                remarks: "כאן ניתן לכתוב הערות",
+              },
+            ],
+          },
+        ],
+      },
+
+      images: {
+        card: {
+          url: checkUrl(req.body.cardUrl),
+          alt: checkReq(req.body.cardAlt),
+        },
+        panorama: {
+          url: checkUrl(req.body.urlPamorama),
+          alt: checkReq(req.body.altPamorama),
+        },
+        before: [
+          {
+            url: checkUrl(req.body.urlBefore),
+            alt: checkReq(req.body.altBefore),
+            description: checkReq(req.body.desBefore),
+          },
+        ],
+        constraction: [
+          {
+            url: checkUrl(req.body.urlConstraction),
+            alt: checkReq(req.body.altConstraction),
+            description: checkReq(req.body.desConstraction),
+          },
+        ],
+        constraction: [
+          {
+            url: checkUrl(req.body.urlConstraction),
+            alt: checkReq(req.body.altConstraction),
+            description: checkReq(req.body.desConstraction),
+          },
+        ],
+        sketches: [
+          {
+            url: checkUrl(req.body.urlSketch),
+            alt: checkReq(req.body.altSketch),
+            description: checkReq(req.body.desSketch),
+            remarks: "כאן ניתן לכתוב הערות",
+          },
+        ],
+        imaging: [
+          {
+            url: checkUrl(req.body.urlImaging),
+            alt: checkReq(req.body.altImaging),
+            description: checkReq(req.body.desImaging),
+            remarks: "כאן ניתן לכתוב הערות",
+          },
+        ],
+        references: [
+          {
+            url: checkUrl(req.body.referenceUrl),
+            alt: checkReq(req.body.referenceAlt),
+            remarks: "כאן ניתן לכתוב הערות",
+          },
+        ],
+        plans: [
+          {
+            url: checkUrl(req.body.urlPlans),
+            alt: checkReq(req.body.altPlans),
+            remarks: "כאן ניתן לכתוב הערות",
+          },
+        ],
+        gallery: [checkUrl(req.body.urlGallery)],
+      },
+      userID: req.body.userID,
+    };
+
+    const { error } = validateProject(project);
+    if (error) return res.status(400).send(error.details[0].message);
+
+    fs.mkdirSync(`public/images/projects/${folder}`);
+
+    project = new Project(project);
+    await project.save();
+
+    user = await User.findOneAndUpdate(
+      { _id: project.userID },
+      {
+        $push: {
+          projects: project._id,
+        },
+      }
+    );
+    await user.save();
+
+    return res.send("הפרויקט נשמר בהצלחה");
+  }
+  return res.status(404).send("You are not authorized to create projects!");
+});
+
 const storage = multer.diskStorage({
   destination: async (req, file, storegePath) => {
     const name = req.body.name;
-    const year = req.body.year;
     const projectFind = await Project.findOne({ name, year });
     if (projectFind)
       return storegePath("פרויקט עם השם הזה כבר נמצא במאגר המידע");
-    const NameLowerCase = name.toLowerCase() + "-" + year.toLowerCase();
-    if (!fs.existsSync(`public/images/projects/${NameLowerCase}`)) {
-      fs.mkdirSync(`public/images/projects/${NameLowerCase}`);
+    const nameLowerCase = name.toLowerCase();
+    if (!fs.existsSync(`public/images/projects/${nameLowerCase}`)) {
+      fs.mkdirSync(`public/images/projects/${nameLowerCase}`);
     }
-    storegePath(null, `public/images/projects/${NameLowerCase}`);
+    storegePath(null, `public/images/projects/${nameLowerCase}`);
   },
   filename: (req, file, setName) => {
     setName(null, file.originalname);
@@ -27,9 +175,8 @@ const storage = multer.diskStorage({
 const storageFromEditFile = multer.diskStorage({
   destination: async (req, file, storegePath) => {
     const name = req.body.name;
-    const year = req.body.year;
-    const NameLowerCase = name.toLowerCase() + "-" + year.toLowerCase();
-    storegePath(null, `public/images/projects/${NameLowerCase}`);
+    const nameLowerCase = name.toLowerCase();
+    storegePath(null, `public/images/projects/${nameLowerCase}`);
   },
   filename: (req, file, setName) => {
     setName(null, file.originalname);
@@ -37,11 +184,11 @@ const storageFromEditFile = multer.diskStorage({
 });
 
 const checkFileType = (file, cb) => {
-  const filetypes = /(jpeg|jpg|png)/;
+  const filetypes = /(jpeg|jpg|png|pdf)/;
   const extname = filetypes.test(path.extname(file.originalname).toLowerCase());
   const mimetype = filetypes.test(file.mimetype);
   if (mimetype && extname) return cb(null, true);
-  return cb("Error: Images Only!");
+  return cb("Error: Images or PDF Only!");
 };
 
 const fileFilter = (req, file, cb) => checkFileType(file, cb);
@@ -58,49 +205,144 @@ const upload = multer({
   fileFilter,
 });
 
-router.post("/", upload.array("images", 20), auth, async (req, res) => {
+router.post("/files", upload.array("images", 25), auth, async (req, res) => {
   if (req.user && req.user.isAdmin) {
-    const name = req.body.name;
-    const year = req.body.year;
-    const folder = name.toLowerCase() + "-" + year.toLowerCase();
+    const { name } = req.body;
+    const folder = name.toLowerCase();
 
     const makeDir = (folder, key) => {
       const image = path.basename(req.body[key]);
       return `/images/projects/${folder}/${image}`;
     };
 
+    const checkName = name => name.replace(/-/g, " ");
+    const checkReq = reqItem => (reqItem ? reqItem : "לא מוגדר");
+    const checkUrl = url =>
+      url ? makeDir(folder, req.body[key]()) : "/images/logo/logo_black.png";
+    const checkPath = path => (path ? path.basename(path) : "לא מוגדר");
+
     let project = {
-      name: req.body.name,
+      name: checkName(req.body.name),
       year: req.body.year,
       size: req.body.size,
       category: req.body.category,
-      description: req.body.description,
-      country: req.body.country,
-      city: req.body.city,
-      cardUrl: makeDir(folder, "cardUrl"),
-      cardAlt: req.body.cardAlt,
-      urlPamorama: makeDir(folder, "urlPamorama"),
-      altPamorama: req.body.altPamorama,
-      urlBefore: makeDir(folder, "urlBefore"),
-      altBefore: req.body.altBefore,
-      desBefore: req.body.desBefore,
-      urlSketch: makeDir(folder, "urlSketch"),
-      altSketch: req.body.altSketch,
-      desSketch: req.body.desSketch,
-      urlImaging: makeDir(folder, "urlImaging"),
-      altImaging: req.body.altImaging,
-      desImaging: req.body.desImaging,
-      urlConstraction: makeDir(folder, "urlConstraction"),
-      altConstraction: req.body.altConstraction,
-      desConstraction: req.body.desConstraction,
-      urlGallery: makeDir(folder, "urlGallery"),
-      altGallery: req.body.altGallery,
+      address: {
+        country: req.body.country,
+        city: req.body.city,
+        street: req.body.street,
+        houseNumber: req.body.houseNumber,
+        zip: checkReq(req.body.zip),
+      },
+      description: checkReq(req.body.description),
+      files: {
+        contracts: [
+          {
+            name: checkPath(req.body.contract),
+            url: checkUrl(req.body.contract),
+            remarks: "כאן ניתן לכתוב הערות",
+          },
+        ],
+        licensing: [
+          {
+            name: checkPath(req.body.licensing),
+            url: checkUrl(req.body.licensing),
+            remarks: "כאן ניתן לכתוב הערות",
+          },
+        ],
+        experts: [
+          {
+            firstName: checkReq(req.body.expertFirstName),
+            lastName: checkReq(req.body.expertLastName),
+            phone: checkReq(req.body.expertPhone),
+            files: [
+              {
+                name: checkPath(req.body.expertFile),
+                url: checkUrl(req.body.expertFile),
+                remarks: "כאן ניתן לכתוב הערות",
+              },
+            ],
+          },
+        ],
+      },
+      images: {
+        card: {
+          url: checkUrl(req.body.cardUrl),
+          alt: checkReq(req.body.cardAlt),
+        },
+        panorama: {
+          url: checkUrl(req.body.urlPamorama),
+          alt: checkReq(req.body.altPamorama),
+        },
+        before: [
+          {
+            url: checkUrl(req.body.urlBefore),
+            alt: checkReq(req.body.altBefore),
+            description: checkReq(req.body.desBefore),
+          },
+        ],
+        constraction: [
+          {
+            url: checkUrl(req.body.urlConstraction),
+            alt: checkReq(req.body.altConstraction),
+            description: checkReq(req.body.desConstraction),
+          },
+        ],
+        constraction: [
+          {
+            url: checkUrl(req.body.urlConstraction),
+            alt: checkReq(req.body.altConstraction),
+            description: checkReq(req.body.desConstraction),
+          },
+        ],
+        sketches: [
+          {
+            url: checkUrl(req.body.urlSketch),
+            alt: checkReq(req.body.altSketch),
+            description: checkReq(req.body.desSketch),
+            remarks: "כאן ניתן לכתוב הערות",
+          },
+        ],
+        imaging: [
+          {
+            url: checkUrl(req.body.urlImaging),
+            alt: checkReq(req.body.altImaging),
+            description: checkReq(req.body.desImaging),
+            remarks: "כאן ניתן לכתוב הערות",
+          },
+        ],
+        references: [
+          {
+            url: checkUrl(req.body.referenceUrl),
+            alt: checkReq(req.body.referenceAlt),
+            remarks: "כאן ניתן לכתוב הערות",
+          },
+        ],
+        plans: [
+          {
+            url: checkUrl(req.body.urlPlans),
+            alt: checkReq(req.body.altPlans),
+            remarks: "כאן ניתן לכתוב הערות",
+          },
+        ],
+        gallery: [checkUrl(req.body.urlGallery)],
+      },
+      userID: req.body.userID,
     };
 
     const { error } = validateProject(project);
     if (error) return res.status(400).send(error.details[0].message);
+
     project = new Project(project);
     await project.save();
+    user = await User.findOneAndUpdate(
+      { _id: project.userID },
+      {
+        $push: {
+          projects: project._id,
+        },
+      }
+    );
+    await user.save();
     return res.send("הפרויקט נשמר בהצלחה");
   }
   return res.send("You are not authorized to create projects");
@@ -155,16 +397,17 @@ router.post("/", upload.array("images", 20), auth, async (req, res) => {
 // });
 
 router.delete("/:id", auth, async (req, res) => {
+  // צריך גם למחוק את התעודת זהות של הפרויקט מהמשתמש שיצר אותו
   if (req.user && req.user.isAdmin) {
     let project = await Project.findOne({ _id: req.params.id });
     if (!project) return res.status(404).send("הפרויקט לא נמצא");
     const name = project.name;
-    const year = project.year;
-    const nameLowerCase = name.toLowerCase() + "-" + year.toLowerCase();
-    const newName = nameLowerCase.replace(/\s/g, "-");
+    const newName = name.replace(/\s/g, "-");
     fs.readdir(`public/images/projects/${newName}`, (err, files) => {
-      for (let file of files) {
-        fs.unlinkSync(`public/images/projects/${newName}/${file}`);
+      if (files.length) {
+        for (let file of files) {
+          fs.unlinkSync(`public/images/projects/${newName}/${file}`);
+        }
       }
       fs.rmdirSync(`public/images/projects/${newName}`);
     });
@@ -173,6 +416,28 @@ router.delete("/:id", auth, async (req, res) => {
   }
   return res.send("You are not authorized to delete projects");
 });
+
+// router.delete("/:id", auth, async (req, res) => {
+//   if (req.user && req.user.isAdmin) {
+//     let project = await Project.findOne({ _id: req.params.id });
+//     if (!project) return res.status(404).send("הפרויקט לא נמצא");
+//     const name = project.name;
+//     const year = project.year;
+//     const nameLowerCase = name.toLowerCase() + "-" + year.toLowerCase();
+//     const newName = nameLowerCase.replace(/\s/g, "-");
+//     fs.readdir(`public/images/projects/${newName}`, (err, files) => {
+//       if (files.length) {
+//         for (let file of files) {
+//           fs.unlinkSync(`public/images/projects/${newName}/${file}`);
+//         }
+//       }
+//       fs.rmdirSync(`public/images/projects/${newName}`);
+//     });
+//     project = await Project.findOneAndRemove({ _id: req.params.id });
+//     return res.send("הפרויקט נמחק");
+//   }
+//   return res.send("You are not authorized to delete projects");
+// });
 
 router.put("/private-area/edit-project-card/:id", auth, async (req, res) => {
   if (req.user && req.user.isAdmin) {
